# openapi.yaml
openapi: 3.0.3
info:
  title: SafeGirl Backend API
  description: |
    Women's safety reporting platform with encrypted reports, blockchain verification,
    non-blocking panic alerts, and report sharing with access control.

    **Key Features:**
    - Phone-only OTP authentication (no passwords)
    - AES-256-GCM encryption for report data
    - IPFS + Polygon blockchain for immutable report records
    - Non-blocking panic alerts with SMS notifications
    - Report access control and sharing
    - Key recovery with PIN backup
  version: 1.0.0
  contact:
    name: SafeGirl Team
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.safegirl.example.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from signup/login verify endpoints
  schemas:
...
    DecryptedReportResponse:
      type: object
      required: [reportId, payload]
      properties:
        reportId:
          type: string
        payload:
          type: object
          description: Decrypted report data
        responses:
          type: array
          items:
            type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
...
    GrantAccessRequest:
      type: object
      required: [reportId, viewerUserId]
      properties:
        reportId:
          type: string
          description: Report to grant access to
        viewerUserId:
          type: string
          description: User to grant access to
        expiresAt:
          type: string
          format: date-time
          nullable: true
          description: Optional expiry time
    GrantAccessResponse:
      type: object
      required: [success, accessId]
      properties:
        success:
          type: boolean
          example: true
        accessId:
          type: string
        message:
          type: string
          example: "Access granted"
    RevokeAccessRequest:
      type: object
      required: [reportId, viewerUserId]
      properties:
        reportId:
          type: string
        viewerUserId:
          type: string
    SharedReportSummary:
      type: object
      required: [reportId, reporterId, sharedAt]
      properties:
        reportId:
          type: string
        reporterId:
          type: string
        sharedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
        isActive:
          type: boolean
    ReportViewersResponse:
      type: object
      required: [reportId, viewers]
      properties:
        reportId:
          type: string
        viewers:
          type: array
          items:
            type: object
            properties:
              viewerId:
                type: string
              grantedAt:
                type: string
                format: date-time
              expiresAt:
                type: string
                format: date-time
                nullable: true
              isActive:
                type: boolean

    # ========== PANIC ALERT SCHEMAS ==========
...
paths:
...
  /api/access/grant:
    post:
      summary: Grant another user access to report
      operationId: grantAccess
      tags: [Access Control]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantAccessRequest'
      responses:
        '200':
          description: Access granted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrantAccessResponse'
        '403':
          description: Not report owner
  /api/access/revoke:
    post:
      summary: Revoke user access to report
      operationId: revokeAccess
      tags: [Access Control]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeAccessRequest'
      responses:
        '200':
          description: Access revoked
  /api/access/shared-with-me:
    get:
      summary: Get reports shared with authenticated user
      operationId: getSharedReports
      tags: [Access Control]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Shared reports list
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: '#/components/schemas/SharedReportSummary'
                  total:
                    type: integer
  /api/access/my-shared-reports:
    get:
      summary: Get reports I've shared with others
      operationId: getMySharedReports
      tags: [Access Control]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: My shared reports
  /api/access/my-report/{reportId}/viewers:
    get:
      summary: Get list of users who have access to my report
      operationId: getReportViewers
      tags: [Access Control]
      security:
        - bearerAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Viewers list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportViewersResponse'
  /api/access/report/{reportId}:
    get:
      summary: View report (must own or have access)
      operationId: viewReport
      tags: [Access Control]
      security:
        - bearerAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Report data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecryptedReportResponse'
        '403':
          description: Access denied

  # ========== PANIC ALERT ENDPOINTS ==========
...